# scheduler: optimize reservation BeforePreFilter performance