### Ⅰ. Describe what this PR does

<!--
- Summarize your change (**mandatory**)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)
-->

Optimize plugin Reservation's BeforePreFilter performance.

```
goos: darwin
goarch: amd64
pkg: github.com/koordinator-sh/koordinator/pkg/scheduler/plugins/reservation
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz

benchmark                                       old ns/op     new ns/op     delta
BenchmarkBeforePrefilterWithMatchedPod-12       7191996       3728866       -48.15%
BenchmarkBeforePrefilterWithMatchedPod-12       6965735       4111975       -40.97%
BenchmarkBeforePrefilterWithMatchedPod-12       7242072       3918338       -45.89%
BenchmarkBeforePrefilterWithMatchedPod-12       7099550       3754856       -47.11%
BenchmarkBeforePrefilterWithMatchedPod-12       7360839       3836024       -47.89%
BenchmarkBeforePrefilterWithMatchedPod-12       7233161       4295326       -40.62%
BenchmarkBeforePrefilterWithMatchedPod-12       7480983       4074013       -45.54%
BenchmarkBeforePrefilterWithMatchedPod-12       7236430       4320509       -40.30%
BenchmarkBeforePrefilterWithMatchedPod-12       6774842       4410080       -34.91%
BenchmarkBeforePrefilterWithMatchedPod-12       6926647       4682762       -32.39%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7741868       2451607       -68.33%
BenchmarkBeforePrefilterWithUnmatchedPod-12     8023967       2744553       -65.80%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7538820       2640237       -64.98%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7633984       2493778       -67.33%
BenchmarkBeforePrefilterWithUnmatchedPod-12     8053347       2841502       -64.72%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7839942       2779620       -64.55%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7826316       2923067       -62.65%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7310364       2461115       -66.33%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7578863       2605870       -65.62%
BenchmarkBeforePrefilterWithUnmatchedPod-12     7700427       2374196       -69.17%


                                   │   old.txt   │               new.txt               │
                                   │   sec/op    │   sec/op     vs base                │
BeforePrefilterWithMatchedPod-12     7.213m ± 4%   4.093m ± 8%  -43.25% (p=0.000 n=10)
BeforePrefilterWithUnmatchedPod-12   7.721m ± 4%   2.623m ± 8%  -66.03% (p=0.000 n=10)
geomean                              7.463m        3.277m       -56.09%

```

### Ⅱ. Does this pull request fix one issue?

fix #1696 

<!--If so, add "fixes #xxxx" below in the next line, for example, fixes #15. Otherwise, add "NONE" -->

### Ⅲ. Describe how to verify it

### Ⅳ. Special notes for reviews

### V. Checklist

- [x] I have written necessary docs and comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `make test`